--- a/zio-query/shared/src/main/scala/zio/query/internal/Result.scala
+++ b/zio-query/shared/src/main/scala/zio/query/internal/Result.scala
@@ -67,6 +67,10 @@ private[query] sealed trait Result[-R, +E, +A] { self =>
       case Fail(e)        => fail(e)
     }
 
+  /**
+    * Zips this result with that result in parallel using the specified
+    * function.
+    */
   final def zipWithPar[R1 <: R, E1 >: E, B, C](that: Result[R1, E1, B])(f: (A, B) => C): Result[R1, E1, C] =
     (self, that) match {
       case (Result.Blocked(br1, c1), Result.Blocked(br2, c2)) => Result.blocked(br1 && br2, c1.zipWithPar(c2)(f))