This PR explores a change to the internal encoding of `ZQuery` to a callback based one:

```scala
final class ZQuery[-R, +E, +A] private (
  private val start: (Result[R, E, A] => UIO[Any]) => ZIO[(R, QueryContext), Nothing, Any]
)
```

This lets us address the issue in ghostdogpr/caliban#577 where we were incurring additional overhead for forking queries that did not actually involve effects in `zipWithPar`. In this encoding, a query can either call the callback synchronously or asynchronously, so we can push the forking into the `fromEffect` constructor and avoid forking when the query is not actually effectual.