@adamgfraser 2020-11-20
Replaced the `Promise` with a `Ref`. Will run all the benchmarks after I finish going through everything. Latest performance on the complex query benchmark is:

```
[info] Benchmark                          Mode  Cnt    Score   Error  Units
[info] GraphQLBenchmarks.complexCaliban  thrpt    5  120.123 ± 2.299  ops/s
[info] GraphQLBenchmarks.complexSangria  thrpt    5  101.097 ± 2.752  ops/s
```