@adamgfraser 2020-11-21
@ghostdogpr Reran the benchmarks for this. Unfortunately, we lose the recent bump in performance we had for the simple benchmark. I think the issue is that now when we have a single effect we still have to fork a fiber whereas we didn't in the previous encoding. Not sure it is possible to avoid that while addressing the complex query case. Do we have a benchmark for more of a mixed "real world query" that we could try this with to see how the two compare on net?