@ghostdogpr 2020-11-22
> @ghostdogpr Reran the benchmarks for this. Unfortunately, we lose the recent bump in performance we had for the simple benchmark. I think the issue is that now when we have a single effect we still have to fork a fiber whereas we didn't in the previous encoding. Not sure it is possible to avoid that while addressing the complex query case. Do we have a benchmark for more of a mixed "real world query" that we could try this with to see how the two compare on net?

Hmm I don't have another one. Real world use cases are very widely split, I guess there might be a lot of simple ones as well as complex ones. People tend to use one extreme or the other. Can you share the benchmark result?