@ghostdogpr 2020-11-22
You could try changing the query to this:
```
val simpleQuery: String =
    """{
          characters{
            name
          }
          character("Amos Burton"){
            name
          }
       }""".stripMargin
```
This would trigger 2 effectual fields instead of 1.