@adamgfraser 2020-11-22
Looks like it is still slightly slower with two effects, which potentially makes sense given we still have to fork two things in both encodings and there is some overhead to doing everything in terms of callbacks, but different is much lower.

Before:

```
[info] Benchmark                             Mode  Cnt      Score     Error  Units
[info] GraphQLBenchmarks.introspectCaliban  thrpt    5   2276.862 ± 128.473  ops/s
[info] GraphQLBenchmarks.introspectSangria  thrpt    5    585.397 ±  63.128  ops/s
[info] GraphQLBenchmarks.simpleCaliban      thrpt    5  17877.155 ± 290.458  ops/s
[info] GraphQLBenchmarks.simpleSangria      thrpt    5   8415.909 ± 252.670  ops/s
```

With callback encoding:

```
[info] Benchmark                             Mode  Cnt      Score     Error  Units
[info] GraphQLBenchmarks.introspectCaliban  thrpt    5   2439.555 ± 193.674  ops/s
[info] GraphQLBenchmarks.introspectSangria  thrpt    5    593.164 ±  54.117  ops/s
[info] GraphQLBenchmarks.simpleCaliban      thrpt    5  15751.370 ± 123.156  ops/s
[info] GraphQLBenchmarks.simpleSangria      thrpt    5   8784.278 ± 184.473  ops/s
```