@ghostdogpr 2020-11-22
Your "before" seems much worse than this https://github.com/ghostdogpr/caliban/issues/577#issuecomment-687717451, I wonder why? In your before/after the degradation is not much.