@ghostdogpr 2020-11-28
I wonder if it's worth it 🤔 I did a bunch of benchmarks before/after, the gains are quite marginal and they benefit the case where you have a huge query full of `ZQuery`. However it's worse for simpler queries and queries using ZIO (`fromEffect`), which I think is the majority.