@adamgfraser 2020-11-28
@ghostdogpr Agreed. This adds significant complexity so if it isn't a clear win in terms of performance I think it is a no. We can always come back to this later if we need to.